+++ lib/js/media_session_web.dart
@@ -1,8 +1,9 @@
 @JS('navigator.mediaSession')
 library media_session_web;
 
+import 'dart:html' as html;
+
 import 'package:js/js.dart';
-import 'media_metadata.dart';
 
 @JS('setActionHandler')
 external void setActionHandler(String action, Function(ActionResult) callback);
@@ -16,7 +17,7 @@ class ActionResult {
   external String get action;
   external double get seekTime;
 
-  external factory ActionResult({String action, double seekTime});
+  external factory ActionResult({String? action, double? seekTime});
 }
 
 @JS()
@@ -26,11 +27,11 @@ class PositionState {
   external double get playbackRate;
   external double get position;
   external factory PositionState({
-    double duration,
-    double playbackRate,
-    double position,
+    double? duration,
+    double? playbackRate,
+    double? position,
   });
 }
 
 @JS('metadata')
-external set metadata(MediaMetadata metadata);
+external set metadata(html.MediaMetadata metadata);